+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Coffee Machine Library Changelog
 
+## []
+### Changed
+- Serial interface output status can now be configured
+
 ## [0.2.0] - 24/02/2021
 ### Changed
 - Simplified example of Adapter class